--- packages/backend/src/models/paramModels.ts
@@ -1,168 +1,0 @@
-import {
-  BlurpleEvent,
-  CanvasInfo,
-  PaletteColor,
-} from "@blurple-canvas-web/types";
-import z from "zod";
-import BadRequestError from "@/errors/BadRequestError";
-
-const CanvasIdParamModel = z.object({
-  canvasId: z.coerce.number().int().positive(),
-});
-
-const EventIdParamModel = z.object({
-  eventId: z.coerce.number().int().positive(),
-});
-
-const ColorIdParamModel = z.object({
-  colorId: z.coerce.number().int().positive(),
-});
-
-const GuildIdParamModel = z.object({
-  guildId: z.string().regex(/^\d+$/, "guildId must be a numeric string"),
-});
-
-export type LeaderboardParamModel = typeof CanvasIdParamModel;
-
-export const LeaderboardQueryModel = z.object({
-  page: z.coerce.number().int().positive().optional(),
-  size: z.coerce.number().int().positive().optional(),
-});
-
-export const PixelHistoryParamModel = z.object({
-  x: z.coerce.number().int().nonnegative(),
-  y: z.coerce.number().int().nonnegative(),
-});
-
-export const FrameGuildIdsQueryModel = z.object({
-  guildIds: z
-    .union([z.string(), z.array(z.string())])
-    .optional()
-    .transform((value) =>
-      value === undefined ? []
-      : Array.isArray(value) ? value
-      : [value],
-    ),
-});
-
-const FrameIdParamModel = z.object({
-  frameId: z.string().regex(/^[0-9a-fA-F]{6}$/),
-});
-
-export const FrameDataParamModel = z
-  .object({
-    name: z.string().min(1).max(100),
-    x0: z.coerce.number().int().nonnegative(),
-    y0: z.coerce.number().int().nonnegative(),
-    x1: z.coerce.number().int().positive(),
-    y1: z.coerce.number().int().positive(),
-  })
-  .superRefine(({ x0, y0, x1, y1 }, ctx) => {
-    if (x0 === x1) {
-      ctx.addIssue({
-        code: "custom",
-        path: ["x1"],
-        message: "x0 must not be equal to x1",
-      });
-    }
-
-    if (y0 === y1) {
-      ctx.addIssue({
-        code: "custom",
-        path: ["y1"],
-        message: "y0 must not be equal to y1",
-      });
-    }
-  });
-export const FrameOwnerParamModel = z.object({
-  ownerId: z.string().regex(/^\d+$/, "ownerId must be a numeric string"),
-  isGuildOwned: z.boolean(),
-});
-
-export interface CanvasIdParam {
-  canvasId: string;
-  [key: string]: string;
-}
-
-export async function parseCanvasId(
-  params: CanvasIdParam,
-): Promise<CanvasInfo["id"]> {
-  const result = await CanvasIdParamModel.safeParseAsync(params);
-  if (!result.success) {
-    throw new BadRequestError(
-      `${params.canvasId} is not a valid canvas ID`,
-      result.error.issues,
-    );
-  }
-
-  return result.data.canvasId;
-}
-
-export interface FrameIdParam {
-  frameId: string;
-  [key: string]: string;
-}
-
-export async function parseFrameId(params: FrameIdParam): Promise<string> {
-  const result = await FrameIdParamModel.safeParseAsync(params);
-  if (!result.success) {
-    throw new BadRequestError(
-      `${params.frameId} is not a valid frame ID`,
-      result.error.issues,
-    );
-  }
-
-  return result.data.frameId;
-}
-
-export interface EventIdParam {
-  eventId: string;
-}
-
-export async function parseEventId(
-  params: EventIdParam,
-): Promise<BlurpleEvent["id"]> {
-  const result = await EventIdParamModel.safeParseAsync(params);
-  if (!result.success) {
-    throw new BadRequestError(
-      `${params.eventId} is not a valid event ID`,
-      result.error.issues,
-    );
-  }
-
-  return result.data.eventId;
-}
-
-export interface ColorIdParam {
-  colorId: string;
-}
-
-export async function parseColorId(
-  params: ColorIdParam,
-): Promise<PaletteColor["id"]> {
-  const result = await ColorIdParamModel.safeParseAsync(params);
-  if (!result.success) {
-    throw new BadRequestError(
-      `${params.colorId} is not a valid color ID`,
-      result.error.issues,
-    );
-  }
-
-  return result.data.colorId;
-}
-
-export interface GuildIdParam {
-  guildId: string;
-}
-
-export async function parseGuildId(params: GuildIdParam): Promise<bigint> {
-  const result = await GuildIdParamModel.safeParseAsync(params);
-  if (!result.success) {
-    throw new BadRequestError(
-      `${params.guildId} is not a valid guild ID`,
-      result.error.issues,
-    );
-  }
-
-  return BigInt(result.data.guildId);
-}
